cli/requirements.txt
rich>=13.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3 